Second commit. Sprint15

+++ transport-catalogue/graph.h
@@ -42,7 +42,7 @@ public:
             transport_catalog_serialize::Edge edge_out;
             edge_out.set_from(static_cast<uint32_t>(edge.from));
             edge_out.set_to(static_cast<uint32_t>(edge.to));
-            edge_out.set_weight(static_cast<uint32_t>(edge.weight));
+            edge_out.set_weight(edge.weight);
             graph.add_edges();
             *graph.mutable_edges(graph.edges_size()-1) = edge_out;
         }
